+++ src/components/index.ts
@@ -13,3 +13,4 @@ export * from '@/components/template/card-pin-input'
 export * from '@/components/template/card-name-input/card-name-input'
 export * from '@/components/template/header/header'
 export * from '@/components/template/card/card'
+export * from '@/components/template/card-type-pick-bottom-sheet/card-type-pick-bottom-sheet'
