+++ src/components/template/card-exp-date-input/card-exp-date-input.css.ts
@@ -5,6 +5,5 @@ export const wrapper = style({
   display: 'grid',
   gridTemplateColumns: '1fr 6px 1fr',
   width: '100%',
-  position: 'relative',
   marginTop: vars.space['4px'],
 })
