+++ src/hooks/use-card-input-state.tsx
@@ -1,26 +1,33 @@
 import { useState } from 'react'
-import { CardInputState } from '@/types/card'
+import { CardType } from '@/constants/card-type.ts'
 
-export const INITIAL_CARD_INPUT_STATE: CardInputState = {
+export interface CardInputState {
+  cardCode: string
+  cardExpDate: string
+  cardName: string
+  cardCVC: string
+  cardPin: string
+  cardType?: CardType
+}
+
+const INITIAL_CARD_STATE: CardInputState = {
   cardCode: '',
   cardExpDate: '',
   cardName: '',
   cardCVC: '',
   cardPin: '',
-  cardNickName: '',
   cardType: undefined,
 }
 
 export const useCardInputState = () => {
-  const [cardInput, setCardInput] = useState<CardInputState>(INITIAL_CARD_INPUT_STATE)
+  const [cardInput, setCardInput] = useState<CardInputState>(INITIAL_CARD_STATE)
 
-  const handleChangeCardInput =
-    (name: keyof CardInputState) => (value: CardInputState[typeof name]) => {
-      setCardInput(prev => ({ ...prev, [name]: value }))
-    }
+  const handleChangeCardInput = (name: keyof CardInputState) => (value: string) => {
+    setCardInput(prev => ({ ...prev, [name]: value }))
+  }
 
-  const resetCardInput = (value?: CardInputState) => {
-    setCardInput(value ?? INITIAL_CARD_INPUT_STATE)
+  const resetCardInput = () => {
+    setCardInput(INITIAL_CARD_STATE)
   }
 
   return { cardInput, setCardInput: handleChangeCardInput, resetCardInput } as const
