--- src/steps/card-register/hooks/use-card-input-state.tsx
@@ -1,31 +1,0 @@
-import { useState } from 'react'
-
-export interface CardInputState {
-  cardCode: string
-  cardExpDate: string
-  cardName: string
-  cardCVC: string
-  cardPin: string
-}
-
-const INITIAL_CARD_STATE: CardInputState = {
-  cardCode: '',
-  cardExpDate: '',
-  cardName: '',
-  cardCVC: '',
-  cardPin: '',
-}
-
-export const useCardInputState = () => {
-  const [cardInput, setCardInput] = useState<CardInputState>(INITIAL_CARD_STATE)
-
-  const handleChangeCardInput = (name: keyof CardInputState) => (value: string) => {
-    setCardInput(prev => ({ ...prev, [name]: value }))
-  }
-
-  const resetCardInput = () => {
-    setCardInput(INITIAL_CARD_STATE)
-  }
-
-  return { cardInput, setCardInput: handleChangeCardInput, resetCardInput } as const
-}
